// src/types/search.ts
export interface SearchResult {
  title: string;        // 장소명 
  link: string;         // 네이버 지도 링크
  category: string;     // 업종 카테고리 
  description: string;  // 장소 설명 
  telephone: string;    // 전화번호
  address: string;      // 지번 주소
  roadAddress: string;  // 도로명 주소
  mapx: string;         // X 좌표 
  mapy: string;         // Y 좌표 
} 